automatically backed up by learn

+++ db/migrate/01_create_users.rb
@@ -4,7 +4,7 @@ class CreateUsers < ActiveRecord::Migration[4.2]
     create_table :users do |u|
       u.string :username
       u.string :password
-      u.float :balance
+      u.decimal :balance
     end
   end
